server/requirements.txt
amqp==5.0.2
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.2
billiard==3.6.3.0
CacheControl==0.12.6
celery==5.0.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cloudinary==1.21.1
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.6
django-celery-results==2.0.0
django-cors-headers==3.2.1
django-extensions==2.2.9
django-filter==2.2.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
flower==0.9.5
gunicorn==19.9.0
html5lib==1.0.1
humanize==3.1.0
idna==2.9
inflection==0.4.0
ipaddr==2.2.0
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
redis==3.5.3
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.2
tornado==6.1
uritemplate==3.0.1
urllib3==1.25.9
uWSGI>2.0.18,<2.1
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==4.1.4
wrapt==1.12.1
